+++ common/api/summary/tree-widget-react.exports.csv
@@ -3,7 +3,6 @@ Release Tag;API Item
 public;areAllModelsVisible(models: string[], viewport: Viewport): boolean
 public;BaseFilterableTreeProps
 public;BaseTreeProps
-internal;CachingElementIdsContainer
 public;CategoriesTreeComponent:
 public;CategoriesTreeComponentProps
 public;CategoriesTreeHeaderButtonProps
@@ -14,22 +13,19 @@ public;CategoryVisibilityHandler
 public;CategoryVisibilityHandlerParams
 public;ClassGroupingOption
 public;ContextMenuItemProps
-internal;createRuleset(props: CreateRulesetProps): Ruleset
-internal;CreateRulesetProps = Omit
-internal;createSearchRuleset(props: CreateSearchRulesetProps): Ruleset
-internal;CreateSearchRulesetProps = Omit
 public;createVisibilityTreeNodeRenderer({ levelOffset, disableRootNodeCollapse, descriptionEnabled, iconsEnabled }: VisibilityTreeNodeRendererProps): (treeNodeProps: TreeNodeRendererProps) => JSX.Element
 public;createVisibilityTreeRenderer({ nodeRendererProps, ...restProps }: VisibilityTreeRendererProps): (treeProps: TreeRendererProps_2) => JSX.Element
-internal;customizeModelsTreeNodeItem(item: Partial
+public;DefaultLabelRenderer({ label, context }: DefaultLabelRendererProps): JSX.Element
+public;DefaultLabelRendererProps
 alpha;ExternalSourcesTree(props: ExternalSourcesTreeProps): JSX.Element
 alpha;ExternalSourcesTreeComponent:
-alpha;ExternalSourcesTreeProps = BaseTreeProps & TreeContextMenuProps
+alpha;ExternalSourcesTreeProps = BaseTreeProps
 public;hideAllCategories(categories: string[], viewport: Viewport): Promise
 public;hideAllModels(models: string[], viewport: Viewport): Promise
 public;IModelContentTree: (props: IModelContentTreeProps) => JSX.Element
 public;IModelContentTreeComponent:
 public;IModelContentTreeComponentProps = Omit
-public;IModelContentTreeProps = BaseTreeProps & TreeContextMenuProps
+public;IModelContentTreeProps = BaseTreeProps
 public;invertAllCategories(categories: CategoryInfo[], viewport: Viewport): Promise
 public;invertAllModels(models: string[], viewport: Viewport): Promise
 public;IVisibilityHandler
@@ -53,12 +49,18 @@ public;showAllCategories(categories: string[], viewport: Viewport): Promise
 public;showAllModels(models: string[], viewport: Viewport): Promise
 internal;SubjectModelIdsCache
 public;toggleModels(models: string[], enable: boolean, viewport: Viewport): Promise
+internal;TREE_NODE_LABEL_RENDERER = "visibility-tree-node-label"
 public;TreeContextMenuItem({ id, children, title, onSelect }: PropsWithChildren
 public;TreeContextMenuItemProps
 public;TreeContextMenuProps
 public;TreeDefinition
-public;TreeRenderer({ contextMenuItems, nodeRenderer, ...restProps }: TreeRendererProps): JSX.Element
-public;TreeRendererBaseProps = TreeContextMenuProps
+internal;TreeNodeLabelRenderer
+public;TreeNodeLabelRendererProps
+public;TreeNodeRendererContext
+internal;TreeNodeRendererContextProvider({ nodeLabelRenderer, node, children }: TreeNodeRendererContextProviderProps): JSX.Element
+internal;TreeNodeRendererContextProviderProps
+public;TreeRenderer({ contextMenuItems, nodeRenderer, nodeLabelRenderer, ...restProps }: TreeRendererProps): JSX.Element
+public;TreeRendererBaseProps = TreeContextMenuProps & TreeNodeRendererContext
 public;TreeRendererProps = TreeRendererProps_2 & TreeRendererBaseProps
 public;TreeWidget
 public;TreeWidgetId = "tree-widget-react:trees"
