+++ packages/itwin/tree-widget/src/components/trees/category-tree/CategoriesTree.tsx
@@ -11,8 +11,7 @@ import { useDisposable } from "@itwin/core-react";
 import { usePresentationTreeNodeLoader } from "@itwin/presentation-components";
 import { Presentation } from "@itwin/presentation-frontend";
 import { TreeWidget } from "../../../TreeWidget";
-import { addCustomTreeNodeItemLabelRenderer } from "../common/TreeNodeRenderer";
-import { combineTreeNodeItemCustomizations } from "../common/Utils";
+import { addCustomTreeNodeItemLabelRenderer, combineTreeNodeItemCustomizations } from "../common/Utils";
 import { VisibilityTreeEventHandler } from "../VisibilityTreeEventHandler";
 import { createVisibilityTreeRenderer, useVisibilityTreeFiltering, VisibilityTreeNoFilteredData } from "../VisibilityTreeRenderer";
 import { CategoryVisibilityHandler } from "./CategoryVisibilityHandler";
